Rename src/components to src/common
