+++ .stylelintignore
@@ -1,0 +1,3 @@
+**/mutations/**
+__generated__
+src/server
