+++ src/common/AppRenderer.js
