+++ src/common/LayoutFooter.js
