+++ src/pages/ErrorPage.js
@@ -10,7 +10,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
-import Link from '../components/Link';
+import Link from '../common/Link';
 
 const color = '#607d8b';
